Move some variables
